organize cache access to math results
by pretending that `bin` takes three arguments, we can create a generic
function pointer and pass it to a new cache wrapper macro.

we can also remove the pure math version of `inner_bic` since it now
takes a function pointer argument.

we also do not need to create functions for each type of cache because
of the new macro, at the expense of a little bit more verbosity.